+++ R/textArea.R
@@ -8,7 +8,8 @@
 #'
 #' If you want to access the component value on request please use \link{valueButton}.
 #'
-#' @param inputId Id of component.
+#' @param inputId Id of component. This is stored as `data-id` attribute to omit automatic binding
+#'   of the element (into textAreaInput).
 #' @param value Initial text area value or value to be updated.
 #' @param label Text area label.
 #' @param width Width of input area.
@@ -20,7 +21,13 @@
 #' @return Nested list of `shiny.tag` objects defining html structure of the component,
 #'   or no value in case of usage of `updateTextArea` method.
 #' @export
-textArea <- function(inputId, value, label, width = "100%", height = "200px", resize = "default", readonly = FALSE) {
+textArea <- function(inputId, value, label, width = "100%", height = "200px", resize = "default", readonly = FALSE, ...) {
+  args <- rlang::dots_list(...)
+
+  if ("id" %in% names(args)) {
+    warning("Assigning id to textarea tag will cause converting it to input binding.")
+  }
+
   read_only <- "readonly"
   if (!readonly) {
     read_only <- NULL
@@ -31,19 +38,22 @@ textArea <- function(inputId, value, label, width = "100%", height = "200px", re
   }
 
   htmltools::attachDependencies(
-    shiny::tagList(
-      shiny::tags$label(`for` = inputId, label),
+    shiny::tags$label(
+      label,
       shiny::tags$textarea(
-        id = inputId,
+        ...,
+        `data-id` = inputId,
         value,
-        class = "form-control",
+        class = "form-control sg_textarea",
         readonly = read_only,
         style = htmltools::css(
           resize = resize,
           width = width,
-          height = height
+          height = height,
+          font.weight = 400
         )
-      )
+      ),
+      style = htmltools::css(width = width)
     ),
     htmltools::htmlDependency(
       name = "textarea",