+++ inst/www/commoninput.js
@@ -8,15 +8,30 @@ var get_binding = function(el) {
 var register_element = function(el) {
   var $element = $(el);
   var $common_wrapper = $element.closest('.sg_common_input');
-  if (!Boolean($common_wrapper)) {
+  if ($common_wrapper.length === 0) {
     return;
   }
   var $common_storage = $('#' + $common_wrapper.data('common_id'));
+  var $ignore_ids = $common_wrapper.data('ignore_ids');
+  var element_binding = get_binding($element);
+  var element_id = element_binding.getId(el);
+  if (Boolean($ignore_ids) && $ignore_ids.split(',').includes(element_id)) {
+    return;
+  }
+  if (Boolean($element.data('ignore'))) {
+    return;
+  }
+  if (!$common_storage.data('el_initialized')) {
+    $common_storage.data('el_initialized', []);
+  }
+  var initialized = $common_storage.data('el_initialized');
+  if (!initialized.includes(element_id)) {
+    initialized.push(element_id);
+  }
+
   if (!$common_storage.data('bound')) {
     return;
   }
-  var element_binding = get_binding($element);
-  var element_id = element_binding.getId(el);
   var registered = $common_storage.data('registered') || [];
   if (!registered.includes(element_id)) {
     element_binding.subscribe($element, $common_storage.data('callback'));
@@ -27,11 +42,29 @@ var register_element = function(el) {
     }
     registered.push(element_id);
   }
-}
+};
 
 $(document).on('shiny:bound', function(event) {
-  register_element(event.target)
-})
+  register_element(event.target);
+});
+
+var assign_subvalues = function(el, inputs, storage_name) {
+      var input_vals = {};
+    var inputs_mapped = $(el).data(storage_name) || [];
+    if (inputs_mapped.length === 0) {
+      return;
+    }
+    inputs.map(function() {
+      var input_element = this;
+      var input_binding = get_binding(input_element);
+      var input_name = input_binding.getId(input_element);
+      if (inputs_mapped.includes(input_name)) {
+        input_vals[input_name] = {"value": input_binding.getValue(input_element), "type": input_binding.getType()};
+      }
+    });
+    $(el).data("value", input_vals);
+    return input_vals;
+};
 
 $.extend(commonInputBinding, {
   find: function(scope) {
@@ -42,26 +75,25 @@ $.extend(commonInputBinding, {
   },
   getValue: function(el) {
     var common_id = el.id;
-    var id_selector = '[data-common_id="' + common_id + '"]'
+    var id_selector = '[data-common_id="' + common_id + '"]';
     var inputs = $(id_selector + '.shiny-bound-input, ' + id_selector + ' .shiny-bound-input');
+    var input_vals;
 
-    var input_vals = {};
-    inputs.map(function() {
-      var input_element = this;
-      var input_binding = get_binding(input_element);
-      var input_name = input_binding.getId(input_element);
-      input_vals[input_name] = input_binding.getValue(input_element);
-    });
+    /* Initial value before subscribe */
+    if (!$(el).data('bound')) {
+      input_vals = assign_subvalues(el, inputs, 'el_initialized');
+      return input_vals;
+    }
 
-    $(el).data("value", input_vals);
-    return(input_vals);
+    input_vals = assign_subvalues(el, inputs, 'registered');
+    return input_vals;
   },
   subscribe: function(el, callback) {
     var common_id = el.id;
     $(el).data('callback', callback);
     $(el).data('registered', []);
     $(el).data('bound', true);
-    var id_selector = '[data-common_id="' + common_id + '"]'
+    var id_selector = '[data-common_id="' + common_id + '"]';
     var inputs = $(id_selector + '.shiny-bound-input, ' + id_selector + ' .shiny-bound-input');
 
     inputs.each(function(index, element) {
@@ -80,4 +112,4 @@ $.extend(commonInputBinding, {
 });
 
 Shiny.inputBindings.register(commonInputBinding, "shiny.commonInputBinding");
-Shiny.inputBindings.setPriority("shiny.commonInputBinding", -1);
+Shiny.inputBindings.setPriority("shiny.commonInputBinding", -2);